Save Meter VAlue

+++ src/main/java/com/mamotec/energycontrolbackend/utils/StringUtils.java
@@ -5,11 +5,11 @@ import java.util.Arrays;
 public class StringUtils {
 
     // String comes in the format of [982,6478]
-    public static int[] toArray(String s) {
+    public static double[] toArray(String s) {
         String cleanedResponse = s.substring(1, s.length() - 1);
 
         return Arrays.stream(cleanedResponse.split(","))
-                .mapToInt(Integer::parseInt)
+                .mapToDouble(Double::parseDouble)
                 .toArray();
     }
